+++ src/ReplicationManagerServer.hpp
@@ -22,9 +22,9 @@ class ReplicationManagerServer
 public:
     ReplicationManagerServer(EntityRegistry& entityRegistry);
     
-    void replicateCreate(uint32_t networkID, uint16_t initialDirtyState);
+    void replicateCreate(uint32_t networkID, uint8_t initialDirtyState);
     void replicateDestroy(uint32_t networkID);
-    void setStateDirty(uint32_t networkID, uint16_t dirtyState);
+    void setStateDirty(uint32_t networkID, uint8_t dirtyState);
     void handleCreateAckd(uint32_t networkID);
     void removeFromReplication(uint32_t networkID);
     void write(OutputMemoryBitStream& ombs, ReplicationManagerTransmissionData* rmtd);
@@ -33,6 +33,6 @@ private:
     EntityRegistry& _entityRegistry;
     std::map<uint32_t, ReplicationCommand> _networkIDToReplicationCommand;
     
-    uint16_t writeCreateAction(OutputMemoryBitStream& ombs, uint32_t networkID, uint16_t dirtyState);
-    uint16_t writeUpdateAction(OutputMemoryBitStream& ombs, uint32_t networkID, uint16_t dirtyState);
+    uint8_t writeCreateAction(OutputMemoryBitStream& ombs, uint32_t networkID, uint8_t dirtyState);
+    uint8_t writeUpdateAction(OutputMemoryBitStream& ombs, uint32_t networkID, uint8_t dirtyState);
 };
